Moved controllers and directives to individual files as panel.js grew too much already and we still need to add a lot of functionality...
