+++ extension/script/devtools/directives/Select2Directive.js
@@ -1,0 +1,16 @@
+'use strict';
+require('angular');
+
+var Site = require('../model/Site');
+
+module.exports = (function () {
+
+    angular.module('KidoScraper').directive('select2', function () {
+        return {
+            restrict: 'A',
+            link: function (scope, el) {
+                $(el).select2();
+            }
+        };
+    })
+})();
